requirements_org.txt
annotated-types==0.6.0
anthropic==0.19.2
anyio==4.2.0
archspec @ file:///croot/archspec_1697725767277/work
asttokens==2.4.1
boltons @ file:///Users/cbousseau/work/recipes/ci_py311/boltons_1677965141748/work
Brotli @ file:///Users/cbousseau/work/recipes/ci_py311/brotli-split_1677936346777/work
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b4nang6w_y/croot/cffi_1700254307954/work
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
conda @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_e1pbeimigh/croot/conda_1701719523512/work
conda-content-trust @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_5324skqvu9/croot/conda-content-trust_1693490622873/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_fc4cx8vjhj/croot/conda-package-handling_1690999937094/work
conda_package_streaming @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_aecpaup22q/croot/conda-package-streaming_1690987978274/work
cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b1p0q5vizk/croot/cryptography_1702070293829/work
decorator==5.1.1
distro==1.9.0
executing==2.0.1
fastapi==0.109.0
filelock==3.13.1
firebase-admin==6.4.0
fsspec==2024.2.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.21.4
idna==3.6
ipython==8.21.0
jedi==0.19.1
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
libmambapy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_59l2npsw_8/croot/mamba-split_1698782625405/work/libmambapy
matplotlib-inline==0.1.6
menuinst @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a2fxmxkg_z/croot/menuinst_1702390297906/work
mistralai==0.0.12
msgpack==1.0.7
openai==1.10.0
orjson==3.9.13
packaging @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6dm6d4jd_t/croot/packaging_1693575176524/work
parso==0.8.3
pexpect==4.9.0
platformdirs @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a8u4fy8k9o/croot/platformdirs_1692205661656/work
pluggy @ file:///Users/cbousseau/work/recipes/ci_py311/pluggy_1677906980825/work
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycosat @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3eg8vdcs6z/croot/pycosat_1696536519213/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b8whqav6qm/croot/pyopenssl_1690223428943/work
pyparsing==3.1.1
PySocks @ file:///Users/cbousseau/work/recipes/ci_py311/pysocks_1677906386870/work
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.1
requests @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_54zi68h2nb/croot/requests_1690400233316/work
rsa==4.9
ruamel.yaml @ file:///Users/cbousseau/work/recipes/ci_py311/ruamel.yaml_1677934845850/work
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.35.1
tokenizers==0.15.2
tqdm==4.66.1
traitlets==5.14.1
truststore @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_42mm7e6j06/croot/truststore_1695244298716/work
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.0.post1
wcwidth==0.2.13
websockets==12.0
zstandard @ file:///Users/cbousseau/work/recipes/ci_py311_2/zstandard_1678996192313/work
